# OpenAPI/api.yaml
...
swagger: '2.0'
info:
  title: Uber API
  description: Move your app forward with the Uber API
  version: "1.0.0"
# the domain of the service
host: api.uber.com
...
schemes:
  - http
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
  /museums:
    get:
      description: Retrieve a list of all Romanian museums available via data.gov
      responses:
        200:
          description: List of museums
          schema:
            type: array
            items:
              $ref: '#/definitions/Museum'
        default:
          description: unexpected error
...
  /artworks:
    get:
      description: Retrieve list of artistic works that satisfy the constrains
      parameters:
      - name: name
        in: query
        type: string
      - name: author
        in: query
        type: string
      - name: museum
        in: query
        type: string
      - name: repositoryId
        in: query
        type: string
      - name: pageSize
        in: query
        type: number
      - name: pageNumber
        in: query
        type: number
      responses:
        200:
          description: Success
          schema: 
            $ref: '#/definitions/Artworks'
        400:
          description: Query parameter of wrong format
        default:
          description: Unexpected Error
...
  /authors:
    get:
      description: Retrieve list of art authors that respect constraints
      parameters:
      - name: name
        in: query
        type: string
      - name: pageSize
        in: query
        type: number
      - name: pageNumber
        in: query
        type: number
      responses:
        200:
          description: Success
          schema: 
            $ref: '#/definitions/Authors'
        400:
          description: Query parameter of wrong format
        default:
          description: Unexpected Error
...
definitions:
  Museum:
    type: object
    properties:
      id:
        type: number
      repositoryId:
        type: string
      repositoryName:
        type: string
      city:
        type: string
  Artworks:
    type: object
    properties:
      numberOfArtworks:
        type: number
      artworks:
        type: array
        items:
          $ref: '#/definitions/Artwork'
  Authors:
    type: object
    properties:
      numberOfAuthors:
        type: number
      authors:
        type: array
        items:
          $ref: '#/definitions/AuthorShort'
  Artwork:
    type: object
    properties:
      id:
        type: number
      title:
        type: string
...
  AuthorShort:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      image:
        type: string
      numberOfArtworks:
        type: number
...